Split services utils hooks
