+++ src/services/check-context-if.js
@@ -1,0 +1,8 @@
+
+import checkContext from './check-context';
+
+export default function (hook, type, methods, label) {
+  if (type && hook.type === type) {
+    checkContext(hook, type, methods, label);
+  }
+}
